pyproject.toml
[tool.poetry]
name = "lestrade"
version = "0.1.0"
description = "Evaluation of multiple databases"
authors = ["Saulo Ferreira <saulo.f.ext@tooseguros.com.br>"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.0"
locust = "^2.12.1"
python-dotenv = "^0.21.0"
cassandra-driver = "3.25.0"
progress = "1.6"
redis = "^4.3.4"
pymongo = "3.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
